claude-code-openai-wrapper/pyproject.toml
[tool.poetry]
name = "claude-code-openai-wrapper"
version = "2.2.0"
description = "OpenAI API-compatible wrapper for Claude Code"
authors = ["Richard Atkinson <richardatk01@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.10.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
sse-starlette = "^2.1.3"
python-multipart = "^0.0.18"
claude-agent-sdk = "^0.1.18"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
requests = "^2.32.0"
openai = "^1.0.0"
pytest-cov = "^7.0.0"
mypy = "^1.14.0"
bandit = "^1.8.0"
safety = "^3.2.0"
hypothesis = "^6.122.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.poetry.scripts]
claude-wrapper = "src.main:run_server"